Move action-types
Rename the directory which contains the action-types to `action-types`
instead of `constants`.  Might as well, eh?
