+++ app/actions/movie.actions.js
@@ -17,7 +17,7 @@ import {
   DISMISSING_MOVIE,
   DISMISSED_MOVIE,
   FAILED_UPDATING_MOVIE,
-} from '../constants/movie.action-types';
+} from '../action-types/movie.action-types';
 
 const loadingMovies = createAction(LOADING_MOVIES);
 const moviesAlreadyLoaded = createAction(MOVIES_ALREADY_LOADED);
