+++ app/reducers/login.reducer.js
@@ -2,7 +2,7 @@ import {
   LOGIN_INITIATED,
   LOGIN_SUCCESS,
   LOGIN_ERROR,
-} from '../constants/login.action-types';
+} from '../action-types/login.action-types';
 
 const initialState = {
   isWaiting: false,
